some more fixes

+++ src/Components/Delete.js
@@ -7,20 +7,22 @@ class Delete extends Component {
 
     handleDelete=(event)=> {
         event.preventDefault();
-        console.log(this.props.student.studentId);
 
-        API.delete('student_cd/delete/'+this.props.student.studentId,
-        {headers:{
-            'Content-Type': 'application/json',
-            Authorization: "Bearer "+ sessionStorage.getItem('token')
-        }}).then(response=>{
-            console.log(response);
-            console.log("Student Deleted");
-            window.location.reload();
-        }).catch(error => {
-            console.log(error);
-            this.setState({ errorMsg: 'Error in recieving Data' });
-        })
+        if(window.confirm("Are you sure???")){
+            API.delete('student_cd/delete/'+this.props.student.studentId,
+            {headers:{
+                'Content-Type': 'application/json',
+                Authorization: "Bearer "+ sessionStorage.getItem('token')
+            }}).then(response=>{
+                console.log(response);
+                console.log("Student Deleted");
+                window.location.reload();
+            }).catch(error => {
+                console.log(error);
+                this.setState({ errorMsg: 'Error in recieving Data' });
+            })
+        }
+        
     }
     render() {
         return (
